reorganize indexes and types data
