+++ server/controllers/search/search.js
@@ -5,7 +5,7 @@ const responses_ = __.require('lib', 'responses')
 const parseResults = require('./lib/parse_results')
 const normalizeResult = require('./lib/normalize_result')
 const boostByPopularity = require('./lib/boost_by_popularity')
-const { possibleTypes } = require('./lib/types')
+const { possibleTypes } = require('./lib/indexes')
 const typeSearch = require('./lib/type_search')
 const sanitize = __.require('lib', 'sanitize/sanitize')
 const Group = __.require('models', 'group')
